gschema: set a more classic style scheme
And by more classic I mean tango.

The problem with the 'solarized-light' style scheme is the background
color, it is not standard for a GTK app.

And the problem with the line numbers' background color that was almost
undistinguishable from the main text's background color is now fixed.

So the 'classic' and 'tango' style schemes work again as expected. The
default value was previously 'classic', but 'tango' has warmer and nicer
colors.