help: tabs: update, improve
For the first modified paragraph: the active tab is no longer larger
than the other tabs, all the tabs have the same length. I think the
active tab is quite distinguishable from the others, with the different
colors/style, no need to explain that part. By testing in the
application, it's self-explanatory.

The rest of the page is a shorter explanation of:
gedit-create-new-file.page + gedit-close-file.page. Maybe that
duplication should be removed. But in gedit-tabs.page there is in
addition the Ctrl+Shift+W shortcut explained.