Remove network monitor, buggy implementation with the GtkInfoBar
The feature was:
When having remote files opened in gedit, an infobar was shown when the
network was no longer available. When the network was again available,
the infobar was hidden.

Problem:
With the way GeditTab is implemented, there can be maximum *one* infobar
shown at a time! So if there is already another infobar shown, it gets
destroyed by the network infobar. Or it gets destroyed when the network
is again available. The problem is that the *other* infobar may be
important in order to bring back the tab in the desired state (editable
text, being able to save the document, etc).

An easy way to reproduce the problem (before this commit):
1. Open an https:// URL.
2. Open a second gedit window and open in it the same https:// URL -> it
shows an infobar to choose whether to keep the document read-only or to
edit anyway.
3. Disconnect the network.
-> the network connectivity infobar has overwritten the infobar to
choose whether to edit the second document.

There are many other situations where this is problematic (see how many
different infobars there are in the gedit codebase).

When GeditTab will be ported to TeplTab, it will be possible to bring
back the feature because with TeplTab it's possible to show several
infobars.