io-error-info-bar: fix fixme in no_backup_saving_error
Reasons to believe that the 'else' block was dead code, impossible to
reach:
- The GError param must match G_IO_ERROR_CANT_CREATE_BACKUP, and that's
when we ask explicitly to create a backup.
- Searching all uses of GEDIT_SETTINGS_CREATE_BACKUP_COPY (or its
equivalent string), it's used only in GeditTab, and GeditTab has
conditional code for the auto-save. When it's an automatic save, we
don't ask to create a backup file (see the comment in gedit-tab.c
function get_initial_save_flags()).
- There is only one call site of
gedit_no_backup_saving_error_info_bar_new(), in GeditTab, where we ask
explicitly or not to create a backup file.

In any case, even if the 'else' block was possible to reach, "backup
file" instead of "temporary backup file" is OK too, it's good enough.

To disambiguate the two cases (if there are actually several cases), a
later commit will add the full error->message to the infobar. That way
the user will have all the infos.