PreferencesDialog: use TeplStyleSchemeChooserWidget to fix several bugs
This changes the UI for choosing the color/style scheme.

TeplStyleSchemeChooserWidget has a simpler implementation, it's a list
with just normal GtkLabels containing the name and description of the
style schemes.

TeplStyleSchemeChooserWidget fixes several bugs:
- Before, with GtkSourceStyleSchemeChooserWidget, the selected color
scheme was not scrolled into view. So if e.g. the tango scheme was
initially selected, when opening the preferences dialog the selected row
was not visible, we needed to scroll down the list to see that tango was
selected.
- The colorful GtkSourceStyleSchemeChooserWidget is actually buggy when
semi-transparent colors are defined in the style scheme. This happens
for the gutter (line numbers) background color, when selecting that
style scheme in the list.