Remove checkupdate plugin
It was built only on Windows, and there are now better ways on Windows
to update software: with the Microsoft Store, or the MSYS2 package
manager.

Also, the less code the better, less things to maintain and to keep in a
working state.

The code is still available in the git history anyway, if really needed
it could be resurrected, moved to the gedit-plugins repo or another
repo.