commands-file: change behavior for remembering FileChooserOpen folder
Before this commit, the FileChooser for opening files was configured to
show the folder of the current document. So depending on the active tab,
opening files with the FileChooser shows different folders.

With this commit, it changes the behavior to simply remember the folder
of the FileChooser, without taking into account the current document.
This requires still a few commits to get things right in that area, to
correctly remember to open files from the Recent list (or not), without
losing the current folder of the 'Save As' FileChooser.