FileChooserDialogOSX: remove it, use the Gtk one on macOS
Ideally gedit should use GtkFileChooserNative in all situations.
GtkFileChooserNative has an implementation for macOS, probably similar
to what Jesse implemented for gedit with GeditFileChooserDialogOSX.

Unfortunately GtkFileChooserNative for macOS currently lacks support for
the new 'choice' API (gtk_file_chooser_add_choice() etc), see:
https://gitlab.gnome.org/GNOME/gtk/-/issues/1983

And when using GtkFileChooserNative, gedit needs to use the 'choice'
API.

In any case, we want to migrate gedit to GtkFileChooserNative
(gradually). Even if we don't enable GtkFileChooserNative everywhere,
the fact that it will be enabled on some platforms means that we will
have the code ready for GtkFileChooserNative.

One thing that I've started to do is to split the Open case from the
Save As case, to not mix the two in the same source file (with common
code in gedit-file-chooser.c), to hopefully have clearer code. Which
means: getting rid of the GeditFileChooserDialog interface (which mixes
the two cases).

It *would* be possible to keep GeditFileChooserDialogOSX and
GeditFileChooserDialog, just removing GeditFileChooserDialogGtk, and
implement the GeditFileChooserOpen/Save interfaces by using
GeditFileChooserDialog. But I would not be able to test it, I don't have
a mac machine.

Long story short, on macOS it's GtkFileChooserDialog that will be used,
like it was the case for gedit pre-2014.