View: remove context menu in the margin to toggle line numbers
That small feature was added in 2009:
commit fb4ced951e10fdd7a28cd95baaf043893e73d7e5

When doing a right click in the left margin/gutter, a context menu
appeared with only one item: to toggle the line numbers visibility.

I don't think it is great to add such options a bit everywhere in the
GUI. The gedit statusbar already contains, nowadays, buttons to toggle
certain settings, including the line numbers visibility. The option is
also of course available in the preferences dialog, where it is the
natural place to change such settings.

Also, when the left gutter contained *only* the line numbers, then doing
a right click to hide the line numbers, the user might not know directly
where to revert the change and display the line numbers again, since the
left gutter is gone.

So, I prefer to remove that small feature. I don't think it was a
well-known feature in any case, since it's a bit hidden behind a right
click at a specific place.